+++ packages/core/src/error/report.ts
@@ -1,4 +1,5 @@
 import { provide } from "inversify-binding-decorators";
+import { Logger } from "../provider/logger/logger-service";
 import { AppError } from "./app-error";
 
 /**
@@ -9,20 +10,25 @@ import { AppError } from "./app-error";
 @provide(Report)
 class Report {
   static stack: string;
+  private logger: Logger;
+
+  constructor() {
+    this.logger = new Logger();
+  }
 
   /**
    * The Error method is responsible for generating a standardized error object,
    * logging the error, and then throwing it for further handling.
-   * The error thrown is of the custom type IAppError, which extends the built-in Error class.
+   * The error thrown is of the custom type AppError, which extends the built-in Error class.
    *
    * @param error - An instance of Error or a string that describes the error.
    * @param statusCode - The HTTP status code associated with the error (default is 500).
    * @param service - The service name associated with the error. If not specified,
    *                  it defaults to the name of the calling function.
    *
-   * @throws An object of the custom type IAppError, which includes details about the error.
+   * @throws An object of the custom type AppError, which includes details about the error.
    */
-  public Error(
+  public error(
     error: Error | string,
     statusCode?: number,
     service?: string,
@@ -35,7 +41,12 @@ class Report {
       appError = new AppError(error, statusCode, service);
     }
 
-    throw appError;
+    this.logger.error(
+      appError.message,
+      appError.service || "service-undefined",
+    );
+
+    return appError;
   }
 }